+++ packages/curd-table/src/components/TableEditItem.vue
@@ -264,12 +264,32 @@ const uploadValue = computed({
     const isArray = Array.isArray(formValue.value);
     const isMultiple = props.option?.config?.multiple;
 
-    if (isArray && isMultiple) {
-      return formValue.value;
-    } else if (!isArray && isMultiple) {
+    if (isMultiple) {
+      // 数组结构数据
+      if (isArray) {
+        // 返回原数组
+        return formValue.value;
+      } else {
+        // 不是数组，判断该值是否为null or undefined
+        if (formValue.value !== null && formValue.value !== undefined) {
+          // 返回数组结构
+          return [formValue.value];
+        } else {
+          // 警告
+          console.warn(
+            `不能将${formValue.value}转为标准上传文件结构`,
+            props.field,
+          );
+          return [];
+        }
+      }
+    } else {
+      // 单数据结构，判断该值是否为null or undefined
       if (formValue.value !== null && formValue.value !== undefined) {
+        // 返回数组结构
         return [formValue.value];
       } else {
+        // 警告
         console.warn(
           `不能将${formValue.value}转为标准上传文件结构`,
           props.field,
@@ -279,6 +299,7 @@ const uploadValue = computed({
     }
   },
   set(v) {
+    console.log(v);
     if (props.option?.config?.multiple) {
       formValue.value = v;
     } else {