+++ packages/ui/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @naive-potato-ui/ui
 
+## 1.0.59
+
+### Patch Changes
+
+- change upload edit
+
+- Updated dependencies []:
+  - @naive-potato-ui/custom-upload@1.0.22
+  - @naive-potato-ui/curd-table@1.0.56
+
 ## 1.0.58
 
 ### Patch Changes